feat: added multimonitor support
Because the original system used a single monitor many of the original
functions had to be reworked to work with the multimonitor setup. A
basic 'wm_t manager' was added so that it managed all monitors.

IMPORTANT: It had a bug with opening neovide so it is very likely that
it has more, still have to figure out how to work with Xephyr on the
multimonitor testing.